requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers>=4.3.0
djangorestframework==3.16.1
idna==3.11
numpy>=1.24.0
python-dotenv==1.2.1
requests==2.32.5
scipy>=1.10.0
sqlparse==0.5.3
urllib3==2.5.0
